Refactor card suit symbol handling and add getSuitSymbol utility function
